lesson 2

+++ lesson 1/utils.h
@@ -1,1 +1,1 @@
-extern const short int sinTable[1440];
+extern const int sinTable[1440];
